refactor: Move global context, hooks, and utils to shared directory
Restructured the codebase to follow the recommended file layout from CLAUDE.md:
- Moved global context files from src/context to src/shared/context
- Moved global hooks from src/hooks to src/shared/hooks
- Moved global utilities from src/utils to src/shared/utils
- Updated all import statements to use the new paths
- Removed duplicate/misplaced type definition files

This change improves code organization by clearly separating shared resources
from feature-specific code, making the codebase more maintainable and scalable.
