+++ src/app/App.tsx
@@ -11,7 +11,7 @@ import posthog from '../helpers/posthog';
 import './App.css';
 import BrandDataLoader from '../features/brand/components/BrandDataLoader';
 import AppRoutes from './AppRoutes';
-import { AppProvider } from '../context/appContext';
+import { AppProvider } from '../shared/context/appContext';
 import { UsersContextProvider } from '../features/users/context/UsersContext';
 import VersionDisplay from '../shared/components/feedback/VersionDisplay';
