+++ src/shared/hooks/useFetchThreadMessages.ts
@@ -1,6 +1,6 @@
 import { useState } from 'react';
-import { getServiceURL } from '../helpers/getServiceURL';
-import { useAuth } from '../features/auth/context/AuthContext';
+import { getServiceURL } from '../../helpers/getServiceURL';
+import { useAuth } from '../../features/auth/context/AuthContext';
 export type Messages = {
   thread_id: string;
   messages: string[];
